CHAT-MYSQL-LANGCHAIN/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-groq
mysql-connector-python
python-dotenv